requirements.txt
# viktor==13.7.0
tensorflow-cpu==2.11.0
